requirements.txt
# requirements.txt

openai
python-dotenv
pydub
python-docx
